pyproject.toml
[project]
name = "chinese-stock-market"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]
